Add Diffux-CI to list of command-line tools
Diffux-CI is a project that @lencioni and I have developed with the goal
of eliminating visual regressions. A few years ago, we co-created
[Diffux](https://github.com/diffux/diffux), a tool to find visual bugs.
Whereas Diffux was all about diffing full pages, Diffux-CI is about
diffing individual components. It is designed to integrate with CI, and
it uses Firefox to take snapshots. Since we introduced Diffux-CI in the
brigade.com codebase six months ago, we've prevented countless visual
regressions and we feel more confident than ever making changes to our
CSS.